requirements.txt
asgiref==3.2.10
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1
djangorestframework==3.12.2
drf-yasg==1.20.0
psycopg2-binary==2.8.6
python-decouple==3.3


